requirements.txt
streamlit==1.45.0
sqlalchemy==2.0.36
anthropic==0.42.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
requests==2.32.3
httplib2==0.22.0
